% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_deseq2.R
\name{call_deseq2}
\alias{call_deseq2}
\title{Performs differential analysis expression using DESeq2}
\usage{
call_deseq2(countData, metadata, useCtrlGenes = FALSE)
}
\arguments{
\item{countData}{Data frame with matrix counts obtained from readTEcounts}

\item{metadata}{Metadata read with read_metadata}

\item{useCtrlGenes}{Boolean for using only genes in estimating DESeq2 size factors (default: FALSE)}
}
\value{
DESeq2 dds object
}
\description{
Performs differential analysis expression using DESeq2
}
