% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTools.R
\name{graphTools}
\alias{graphTools}
\title{Functions to plot DESeq2 results}
\usage{
graphTools(
  res,
  maxpadj,
  minlfc,
  device = "png",
  output_folder = ".",
  width = 7,
  height = 7,
  plot.title = NULL
)
}
\arguments{
\item{res}{}

\item{maxpadj}{}

\item{minlfc}{}

\item{plot.title}{}
}
\description{
Functions to plot DESeq2 results
}
