% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_counts.R
\name{norm_counts}
\alias{norm_counts}
\title{Get normalized counts from DESeq2 object}
\usage{
norm_counts(dds)
}
\arguments{
\item{dds}{DEseq2 object}
}
\value{
Data frame with normalized counts by samples
}
\description{
Get normalized counts from DESeq2 object
}
