% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Reads metadata file with sample information}
\usage{
read_metadata(datafile)
}
\arguments{
\item{datafile}{A tab separated file with four columns and a header. Structure
must follow this example:

File         Sample    Group     Condition
WT1.cntTable    WT1        WT       Control
....
KO1.cntTable    KO1        KO       Treat

"Control" and "Treat" labels are required for DESeq2 analysis.}
}
\value{
Data frame with metadata
}
\description{
Reads metadata file with sample information
}
\examples{
datafile <- "my.path/data.csv"
metadata <- read_metadata(datafile)
}
