% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTEcounts.R
\name{readTEcounts}
\alias{readTEcounts}
\title{Read TE count data and merge in a data frame by sample}
\usage{
readTEcounts(metadata, folder)
}
\arguments{
\item{metadata}{Data frame with sample information, previously read with
function read_metadata.}

\item{folder}{Folder for count files listed in first column}
}
\value{
Data frame with counts by features in rows and samples in columns
}
\description{
Read TE count data and merge in a data frame by sample
}
\examples{
datafile <- "my.path/data.csv"
metadata <- read_metadata(datafile)
folder <- "data/"
df <- readTEexpress(metadata, folder)
}
