% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_deseq2.R
\name{results_deseq2}
\alias{results_deseq2}
\title{Get results after DESeq2 analysis}
\usage{
results_deseq2(dds, shrinklog2FC = FALSE)
}
\arguments{
\item{dds}{Object from call_deseq2 function}

\item{shrinklog2FC}{Boolean for applying log2FC shrinking in DESEq2 (default: FALSE)}
}
\value{
Data frame with DESeq2 results
}
\description{
Get results after DESeq2 analysis
}
