% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_DEA.R
\name{TE_DEA}
\alias{TE_DEA}
\title{Compute differential expression analysis for transposable elements}
\usage{
TE_DEA(metafile, folder, useCtrlGenes = FALSE, shrinklog2FC = FALSE)
}
\arguments{
\item{metafile}{}

\item{folder}{}

\item{useCtrlGenes}{}

\item{shrinklog2FC}{}
}
\value{
Table of results
}
\description{
Compute differential expression analysis for transposable elements
}
